[AMD-SMI] Fix additional trivial correctness and safety bugs from code hygiene audit
Fixes bugs of the same classes addressed in commit 562d704b3b, found in
code added since that commit:

- rocm_smi_kfd.cc: stoi → stoull for GPU ID stored in uint64_t (same
  bug as previously fixed at line 435)
- amd_smi_utils.cc: Guard pclose(NULL) UB in shared_ptr<FILE> deleter
- rocm_smi.cc: Fix refcount/fd leak on 3 error paths in
  rsmi_event_notification_init() — refcount was incremented but never
  decremented on ioctl failure, invalid anon_fd, or fdopen failure
- rocm_smi.cc: Add %255s/%255[^\n] width specifiers to sscanf calls
- Replace &vec[0] with vec.data() across production code, tests, and
  examples to eliminate UB when vectors are empty (31 locations)
- amd_smi.cc: Convert non-standard VLA to std::vector

+++ projects/amdsmi/rocm_smi/src/rocm_smi_kfd.cc
@@ -668,7 +668,7 @@ int GetProcessGPUs(uint32_t pid, std::unordered_set<uint64_t>* gpu_set) {
 
       uint64_t val;
       try {
-        val = static_cast<uint64_t>(std::stoi(tmp));
+        val = std::stoull(tmp);
       } catch (...) {
         std::cerr << "Error; read invalid data: " << tmp << " from " << q_gpu_id_str << std::endl;
         closedir(queues_dir_hd);
